FLYW-57 move: 헤더 공통 resource common 폴더 이동 (#23)
