+++ src/main/webapp/WEB-INF/views/search.jsp
@@ -9,9 +9,9 @@
 
     <title>검색 결과 - Flyway</title>
 
-    <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/search/css/variables.css">
-    <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/search/css/global.css">
-    <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/search/css/header.css">
+    <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/common/css/variables.css">
+    <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/common/css/global.css">
+    <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/common/css/header.css">
     <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/search/css/search.css?v=<%= System.currentTimeMillis() %>">
     <link rel="stylesheet" href="${pageContext.request.contextPath}/resources/search/css/flights.css?v=<%= System.currentTimeMillis() %>">
 </head>
@@ -21,7 +21,7 @@
 <header class="header">
     <div class="header__content">
         <div class="header__logo">
-            <img src="${pageContext.request.contextPath}/resources/search/img/logo.svg" alt="Flyway" />
+            <img src="${pageContext.request.contextPath}/resources/common/img/logo.svg" alt="Flyway" />
         </div>
         <nav class="header__menu">
             <a href="${pageContext.request.contextPath}/login" class="header__menu-item">로그인</a>
