+++ src/main/webapp/resources/auth/home.css
@@ -4,6 +4,7 @@
     padding: 42px 40px;
     border-radius: 16px;
     width: 420px;
+    max-width: 100%;
     box-shadow: 0 20px 40px rgba(0, 0, 0, 0.08),
     0 2px 8px rgba(0, 0, 0, 0.05);
     text-align: center;
@@ -90,13 +91,7 @@ a {
 body {
     margin: 0;
     min-height: 100vh;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    background: linear-gradient(180deg, #f4f6fb, #eef1f6);
-    font-family: 'Malgun Gothic', -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
+    background: var(--color-bg-main);
 }
 
 button {
@@ -107,4 +102,40 @@ button {
     font: inherit;
     color: inherit;
     cursor: pointer;
+}
+
+.home-main {
+    max-width: 1100px;
+    margin: 0 auto;
+    padding: 0 20px 60px;
+    position: relative;
+    top: -80px;
+    display: flex;
+    justify-content: center;
+}
+
+.header__menu-form {
+    margin: 0;
+}
+
+.header__menu-button {
+    background: none;
+    border: none;
+    padding: 0;
+    font: inherit;
+    color: inherit;
+    cursor: pointer;
+}
+
+.header__menu-item--cta {
+    padding: 8px 16px;
+    border-radius: 999px;
+    background: var(--color-primary);
+    color: var(--color-white);
+    font-weight: var(--font-weight-semibold);
+}
+
+.header__menu-item--cta:hover {
+    color: var(--color-white);
+    opacity: 0.9;
 }
