+++ src/main/webapp/resources/common/css/global.css
