+++ src/main/webapp/resources/common/css/variables.css
