+++ src/main/webapp/resources/common/js/authFetch.js
